renamed send/receiver for c samples
