+++ lang/c/samples/CMakeLists.txt
@@ -21,8 +21,8 @@ cmake_minimum_required(VERSION 3.10)
 # c samples
 # --------------------------------------------------------
 # pubsub
-add_subdirectory(pubsub/string/hello_snd)
-add_subdirectory(pubsub/string/hello_rec)
+add_subdirectory(pubsub/string/hello_send)
+add_subdirectory(pubsub/string/hello_receive)
 
 #service
 add_subdirectory(services/minimal_client_c)
